unnest Schema, add @xgovformbuilder/model to runner deps
